Finish top, information, and quote sections
HTML and CSS of about 3/4 of the landing page. Forgot to commit after
each section, will have to remember to do that for the rest of this
project and future ones.

The top section includes a navbar with a logo on the left side as well
as a hero section that acts as an attention grabber for some product.
The information section allows different aspects of that product to be
displayed and emphasized to visitors and the quote section allows
visitors to hear a first hand comment from a previous customer.